requirements.txt
# AWS Monitor Dependencies
# Python 3.11+ required

# LLM Integration
litellm>=1.50.0

# AWS SDK
boto3>=1.35.0
aiobotocore>=2.15.0

# Web Framework (for API)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Async HTTP
aiohttp>=3.10.0
httpx>=0.27.0

# Configuration
PyYAML>=6.0.2
python-dotenv>=1.0.0
pydantic>=2.9.0

# Utilities
tenacity>=9.0.0  # Retry logic
croniter>=3.0.0  # Cron scheduling

# Logging and Monitoring
structlog>=24.4.0
rich>=13.9.0  # Pretty console output

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
